removed redundat class
